backend/requirements.txt
google-adk>=0.0.5
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
# react>=18.2.0      # frontend
# tailwindcss>=3.3.5 # frontend
supabase>=2.15.1   # ← NEW
pydantic>=2.4.2
websockets>=12.0.0
python-socketio>=5.10.0
litellm>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.5.0
# file-processing deps
pandas>=2.0.0
openpyxl>=3.1.2
pypdf2>=3.0.0
python-multipart>=0.0.6
# async driver for supabase-py (opcional pero recomendado)
asyncpg>=0.29.0
